# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'API документашка'
  version: 1.0.0
paths:
...
  /api/orders:
    get:
      tags:
        - Order
      summary: 'INDEX - Список всех заказов'
      operationId: 2cd7e03021e72d9d799c12dfe179c865
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      properties:
                        id:
                          type: integer
                          example: 1
                        total:
                          type: integer
                          example: 46200
                        name:
                          type: string
                          example: 'Семён Семёныч'
                        user_id:
                          type: integer
                          example: 1
                        phone:
                          type: string
                          example: '+79999999999'
                        email:
                          type: string
                          example: example@mail.org
                        address:
                          type: string
                          example: 'ул. Кукурузная, д. 35'
                        comment:
                          type: string
                          example: Побыстрее!
                        status:
                          type: string
                          example: 'Создан, ожидает подтверждения.'
                      type: object
                  links:
                    type: obj
                    example: '{...}'
                  meta:
                    type: obj
                    example: '{...}'
                type: object
        404:
          description: 'Нет прав = не знаешь о существовании страницы'
      security:
        -
          jwt: []
    post:
      tags:
        - Order
      summary: 'STORE - Создание заказа'
      operationId: d690d6834318e9b064df441e854de8b9
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  example: 'Василий Иваныч'
                phone:
                  type: string
                  example: '+70123456789'
                email:
                  type: string
                  example: ivanich@mail.org
                address:
                  type: string
                  example: 'ул. Ленина, д. 1'
                comment:
                  type: string
                  example: 'Хочу скидку побольше!'
                products:
                  type: array
                  items:
                    properties:
                      product_id:
                        type: integer
                        example: 1
                      count:
                        type: integer
                        example: 3
                    type: object
              type: object
      responses:
        201:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Заказ успешно создан.'
                  order:
                    type: obj
                    example: '{ id, ... }'
                  errors:
                    type: obj
                    example: '{...}'
                type: object
        422:
          description: 'Валидация не пройдена'
  '/api/orders/{order}/status':
    patch:
      tags:
        - Order
      summary: 'changeStatus - Изменение статуса заказа'
      operationId: 25047fcd73201f1d4ba09153eaf5cef3
      parameters:
        -
          name: order
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                status:
                  type: integer
                  example: 2
              type: object
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Статус заказа #1 успешно изменён'
                type: object
        400:
          description: 'Неверный id статуса'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Указан id несуществующего статуса.'
                type: object
        404:
          description: 'Нет прав = не знаешь о существовании страницы'
      security:
        -
          jwt: []
  '/api/orders/{order}':
    get:
      tags:
        - Order
      summary: 'SHOW - Отображение заказа'
      operationId: f090f23952a43521af6f1d133826bc37
      parameters:
        -
          name: order
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data:
                    properties:
                      id:
                        type: integer
                        example: 1
                      total:
                        type: integer
                        example: 46200
                      name:
                        type: string
                        example: 'Василий Иваныч'
                      phone:
                        type: string
                        example: '+70123456789'
                      email:
                        type: string
                        example: example@mail.org
                      address:
                        type: string
                        example: 'ул. Ленина, д. 1'
                      comment:
                        type: string
                        example: Поживее!
                      status:
                        type: string
                        example: 'Подтверждён, ожидает оплаты.'
                      created_at:
                        type: string
                        example: '2024-04-09T20:08:18.000000Z'
                      updated_at:
                        type: string
                        example: '2024-04-09T16:08:40.000000Z'
                      products:
                        type: array
                        items:
                          properties:
                            id:
                              type: integer
                              example: 1
                            name:
                              type: string
                              example: 'JBL pariatur'
                            slug:
                              type: string
                              example: jbl-pariatur
                            image:
                              type: string
                              example: 'http://localhost:8000/storage/photos/products/overhead/wired/0.jpg'
                            count:
                              type: integer
                              example: 2
                            price:
                              type: integer
                              example: 23300
                          type: object
                    type: object
                type: object
        404:
          description: 'Нет прав = не знаешь о существовании страницы'
      security:
        -
          jwt: []
    delete:
      tags:
        - Order
      summary: 'DESTROY - Удаление заказа'
      operationId: 173211a50d204e6c949d9b18baa95efb
      parameters:
        -
          name: order
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Заказ удалён.'
                  order:
                    properties:
                      id:
                        type: integer
                        example: 1
                      user_id:
                        type: integer
                        example: 1
                      status_id:
                        type: integer
                        example: 1
                      total:
                        type: integer
                        example: 46200
                      name:
                        type: string
                        example: 'Василий Иваныч'
                      phone:
                        type: string
                        example: '+70123456789'
                      email:
                        type: string
                        example: ivanich@mail.org
                      address:
                        type: string
                        example: 'ул. Ленина, д. 1'
                      comment:
                        type: string
                        example: Поживее!
                      created_at:
                        type: string
                        example: '2024-04-09T20:08:18.000000Z'
                      updated_at:
                        type: string
                        example: '2024-04-09T16:08:40.000000Z'
                    type: object
                type: object
        404:
          description: 'Нет прав = не знаешь о существовании страницы'
      security:
        -
          jwt: []
  '/api/users/{id}/orders':
    get:
      tags:
        - Order
      summary: 'showByUserId - Список заказов пользователя'
      operationId: e883720a0185d1975acb17228d8e5096
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      properties:
                        id:
                          type: integer
                          example: 1
                        total:
                          type: integer
                          example: 46200
                        address:
                          type: string
                          example: 'ул. Ленина, д. 1'
                        comment:
                          type: string
                          example: Поживее!
                        status:
                          type: string
                          example: 'Подтверждён, ожидает оплаты.'
                        created_at:
                          type: string
                          example: '2024-04-09T20:08:18.000000Z'
                        updated_at:
                          type: string
                          example: '2024-04-09T16:08:40.000000Z'
                      type: object
                  links:
                    type: obj
                    example: '{...}'
                  meta:
                    type: obj
                    example: '{...}'
                type: object
        404:
          description: 'Нет прав = не знаешь о существовании страницы'
      security:
        -
          jwt: []
...
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer